+++ chart/scripts/twistlock-init.sh
@@ -10,8 +10,8 @@
 # shellcheck disable=SC1091
 
 set -e
-# set -v
-# set -x
+#set -v
+#set -x
 
 # Import common environment variables and functions
 MYDIR="$(dirname "$(readlink -f "$0")")"